backend/requirements.txt
pandas
python-dotenv
langchain
langchain-groq
langchain-openai
flask
